Add test that shows that there are memory problems.
The current design requires the user to keep the document node around as long as they are accessing any
of the nodes from below that. This doesn't fit with Swift's general notion of only holding on to objects
that you still want. I think that people will get confused because they may have pulled out and saved
references to the nodes that they care about, and just let the document be released because they
don't need to access it anymore.

For example, if you make up an array of the nodes that you want to deal with, you might assume that
you no longer need to hold onto the document. But, that would be wrong, as this added test shows.