Fix memory allocation problem by relying on Swift reference counting
Rather than trying to handle memory deallocation via flags (which wasn't working),
change to relying on Swift reference counting by specifying which object
actually should be considered to "owning" the cmarkNode memory, and have
other objects reference that via normal Swift references. That way,
as long as any of the objects who need the cmarkNode memory are alive,
the original document node will still be alive and holding on to the cmarkNode
memory. Once all references to that document node are gone, then the
cmarkNode memory can be released.