CmNode memory fix (#29)
* Add test that shows that there are memory problems.

The current design requires the user to keep the document node around as long as they are accessing any
of the nodes from below that. This doesn't fit with Swift's general notion of only holding on to objects
that you still want. I think that people will get confused because they may have pulled out and saved
references to the nodes that they care about, and just let the document be released because they
don't need to access it anymore.

For example, if you make up an array of the nodes that you want to deal with, you might assume that
you no longer need to hold onto the document. But, that would be wrong, as this added test shows.

* Fix memory allocation problem by relying on Swift reference counting

Rather than trying to handle memory deallocation via flags (which wasn't working),
change to relying on Swift reference counting by specifying which object
actually should be considered to "owning" the cmarkNode memory, and have
other objects reference that via normal Swift references. That way,
as long as any of the objects who need the cmarkNode memory are alive,
the original document node will still be alive and holding on to the cmarkNode
memory. Once all references to that document node are gone, then the
cmarkNode memory can be released.

* Add more tests for memory references.